som fixes
